chore(components): rename logo components
