+++ packages/components/src/components/logos/CoinLogo/CoinLogo.stories.tsx
