+++ packages/components/src/components/logos/CoinLogo/CoinLogo.tsx
