+++ packages/components/src/components/logos/TrezorLogo/TrezorLogo.stories.tsx
