+++ packages/components/src/components/logos/TrezorLogo/TrezorLogo.tsx
