+++ packages/components/src/components/logos/logos.stories.tsx
