+++ packages/components/src/index.ts
@@ -33,8 +33,8 @@ export * from './components/others/SecurityCard';
 export * from './components/prompts/ConfirmOnDevice';
 export * from './components/Modal';
 export * from './components/Modal/Backdrop';
-export * from './components/logos/CoinLogo';
-export * from './components/logos/TrezorLogo';
+export * from './components/logos/CoinLogo/CoinLogo';
+export * from './components/logos/TrezorLogo/TrezorLogo';
 export * from './components/loaders/Loader';
 export * from './components/loaders/Progress';
 export * from './components/loaders/FluidSpinner';
